+++ samples/embeddings/csharp-ooproc/Embeddings/Program.cs
@@ -1,12 +1,10 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using System.Text.Json.Serialization;
-using System.Text.Json;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Azure.Functions.Worker.Extensions.OpenAI.Embedding;
+
 
 var host = new HostBuilder()
     .ConfigureFunctionsWebApplication()