+++ src/WebJobs.Extensions.OpenAI/Embedding/EmbeddingsConverter.cs
@@ -1,15 +1,13 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using System.ClientModel.Primitives;
 using System.Text.Json;
 using Azure;
 using Azure.AI.OpenAI;
 using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
-using WebJobs.Extensions.OpenAI;
 
-namespace Microsoft.Azure.WebJobs.Extensions.OpenAI;
+
+namespace Microsoft.Azure.WebJobs.Extensions.OpenAI.Embedding;
 
 class EmbeddingsConverter :
     IAsyncConverter<EmbeddingsAttribute, EmbeddingsContext>,
@@ -17,6 +15,11 @@ class EmbeddingsConverter :
 {
     readonly OpenAIClient openAIClient;
     readonly ILogger logger;
+    // Note: we need this converter as Azure.AI.OpenAI does not support System.Text.Json serialization since their constructors are internal
+    static readonly JsonSerializerOptions options = new()
+    {
+        Converters = { new EmbeddingsJsonConverter() }
+    };
 
     public EmbeddingsConverter(OpenAIClient openAIClient, ILoggerFactory loggerFactory)
     {
@@ -36,12 +39,6 @@ class EmbeddingsConverter :
         CancellationToken cancellationToken)
     {
         EmbeddingsContext response = await this.ConvertCoreAsync(input, cancellationToken);
-
-        // Note: we need this converter as Azure.AI.OpenAI does not support System.Text.Json serialization since their constructors are internal
-        var options = new JsonSerializerOptions
-        {
-            Converters = { new EmbeddingsJsonConverter() }
-        };
         return JsonSerializer.Serialize(response, options);
     }