+++ src/WebJobs.Extensions.OpenAI/Embedding/EmbeddingsJsonConverter.cs
@@ -4,18 +4,16 @@
 using System.ClientModel.Primitives;
 using System.Text.Json;
 using System.Text.Json.Serialization;
-using Azure;
 using Azure.AI.OpenAI;
-using Azure.Core;
-using Microsoft.Azure.WebJobs.Extensions.OpenAI;
 
-namespace WebJobs.Extensions.OpenAI;
+namespace Microsoft.Azure.WebJobs.Extensions.OpenAI.Embedding;
 
 /// <summary>
 /// Embeddings JSON converter needed to serialize the EmbeddingsContext object.
 /// </summary>
-internal class EmbeddingsJsonConverter : JsonConverter<EmbeddingsContext>
+class EmbeddingsJsonConverter : JsonConverter<EmbeddingsContext>
 {
+    static readonly ModelReaderWriterOptions modelReaderWriterOptions = new("J");
     public override EmbeddingsContext Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
     {
         throw new NotImplementedException();
@@ -25,10 +23,10 @@ internal class EmbeddingsJsonConverter : JsonConverter<EmbeddingsContext>
     {
         writer.WriteStartObject();
         writer.WritePropertyName("request"u8);
-        ((IJsonModel<EmbeddingsOptions>)value.Request).Write(writer, new ModelReaderWriterOptions("J"));
+        ((IJsonModel<EmbeddingsOptions>)value.Request).Write(writer, modelReaderWriterOptions);
 
         writer.WritePropertyName("response"u8);
-        ((IJsonModel<Embeddings>)value.Response).Write(writer, new ModelReaderWriterOptions("J"));
+        ((IJsonModel<Embeddings>)value.Response).Write(writer, modelReaderWriterOptions);
 
         writer.WritePropertyName("count"u8);
         writer.WriteNumberValue(value.Count);