+++ src/WebJobs.Extensions.OpenAI/OpenAIExtension.cs
@@ -5,6 +5,7 @@ using Azure;
 using Azure.AI.OpenAI;
 using Microsoft.Azure.WebJobs.Description;
 using Microsoft.Azure.WebJobs.Extensions.OpenAI.Assistants;
+using Microsoft.Azure.WebJobs.Extensions.OpenAI.Embedding;
 using Microsoft.Azure.WebJobs.Extensions.OpenAI.Models;
 using Microsoft.Azure.WebJobs.Extensions.OpenAI.Search;
 using Microsoft.Azure.WebJobs.Host.Config;