+++ samples/chat/nodejs/src/functions/app.ts
@@ -45,7 +45,8 @@ app.http('GetChatState', {
 
 const chatBotPostOutput = output.generic({
     type: 'assistantPost',
-    id: '{chatID}'
+    id: '{chatID}',
+    model: '%CHAT_MODEL_DEPLOYMENT_NAME%'
 })
 app.http('PostUserResponse', {
     methods: ['POST'],