+++ src/WebJobs.Extensions.OpenAI/Search/SemanticSearchAttribute.cs
@@ -70,7 +70,7 @@ public class SemanticSearchAttribute : Attribute
 
     /// <summary>
     /// Gets or sets the ID of the model to use for embeddings.
-    /// The default value is "text-embedding-ada-002".
+    /// The default value is "text-embedding-3-small".
     /// </summary>
     /// <remarks>
     /// This property supports binding expressions.