+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.13.0 - Unreleased
+
+### Breaking changes
+
+- Changed default embeddings model from `text-embedding-ada-002` to `text-embedding-3-small`
+
 ## v0.12.0 - 2024/03/01
 
 ### Changes